chore: rename org.holochain.androidserviceclient.client -> org.holochain.androidserviceclient.service to avoid ktlint complaint about package names with underscores
