+++ CHANGELOG.md
@@ -26,4 +26,4 @@
 - Display notice on consumer app launch when unable to connect to HolochainService.
 - Refactor setupApp implementation by moving core logic into the `holochain-service-runtime` crate, rename `tauri-plugin-holochain-service-consumer` command from `setupApp` to `connectSetupApp`.
 - Split HolochainService IPC binders into "admin" and "app" binders. Restrict admin binder calls to only the same android package as the service. Restrict app binder calls to only authorized package + happ pairs. Currently new package + happ pairs are authorized automatically.
-- Extract HolochainService into standalone kotlin library `org.holochain.androidserviceruntime.holochain_service` to simplify testing. Run tests in CI.
+- Extract HolochainService into standalone kotlin library `org.holochain.androidserviceruntime.service` to simplify testing. Run tests in CI.
