+++ crates/holochain-conductor-runtime-ffi/uniffi.toml
@@ -1,4 +1,4 @@
 [bindings.kotlin]
-package_name = "org.holochain.androidserviceruntime.holochain_service"
+package_name = "org.holochain.androidserviceruntime.service"
 cdylib_name = "holochain_conductor_runtime_ffi"
 android = true
