+++ crates/tauri-plugin-holochain-service/android/build.gradle.kts
@@ -59,7 +59,7 @@ java {
 dependencies {
     // Subprojects
     implementation(project(":tauri-android"))
-    implementation("org.holochain.androidserviceruntime:holochain-service:0.0.11")
+    implementation("org.holochain.androidserviceruntime:service:0.0.11")
     implementation("org.holochain.androidserviceruntime:client:0.0.11")
     // Kotlin
     implementation("androidx.core:core-ktx:1.7.0")
