+++ crates/tauri-plugin-holochain-service/android/src/main/java/com/plugin/holochain_service/HolochainServicePlugin.kt
@@ -18,7 +18,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.launch
-import org.holochain.androidserviceruntime.holochain_service.HolochainService
+import org.holochain.androidserviceruntime.service.HolochainService
 import org.holochain.androidserviceruntime.client.HolochainServiceAdminClient
 import org.holochain.androidserviceruntime.client.toJSONArrayString
 import org.holochain.androidserviceruntime.client.toJSONObjectString
