+++ libraries/service/.gitignore
