+++ libraries/service/gradle.properties.example
