+++ libraries/service/settings.gradle
