+++ libraries/service/src/main/java/org/holochain/androidserviceruntime/service/Exceptions.kt
@@ -1,4 +1,4 @@
-package org.holochain.androidserviceruntime.holochain_service
+package org.holochain.androidserviceruntime.service
 
 class UnauthorizedException(
     message: String = "Unauthorized",
