+++ libraries/service/src/main/res/drawable/notification_icon_sm.png
